Use pytest cli option to choose simulator in general conv tests
