--- myhdl/test/bugs/GHDL.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import
-from myhdl.conversion import verify
-from myhdl.conversion import analyze
-
-verify.simulator = "GHDL"
-analyze.simulator = "GHDL"
