+++ myhdl/test/bugs/Makefile
@@ -2,20 +2,20 @@ all: vlog vcom
 
 
 vlog:
-	py.test vlog.py test_*.py
+	py.test --sim vlog
 
 vcom:
-	py.test vcom.py test_*.py
+	py.test --sim vcom
 
 
 GHDL:
-	py.test GHDL.py test_*.py
+	py.test --sim GHDL
 
 icarus:
-	py.test icarus.py test_*.py
+	py.test --sim icarus
 
 cver:
-	py.test cver.py test_*.py
+	py.test --sim cver
 
 clean:
 	- rm *.o *.out *.v *.vhd *.pyc *~ *.vcd* *.log *_ghdl
