--- myhdl/test/bugs/vlog.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import
-from myhdl.conversion import verify, analyze
-
-verify.simulator = analyze.simulator = "vlog"
