--- myhdl/test/conversion/toVHDL/GHDL.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import
-from myhdl.conversion import verify
-
-verify.simulator = "GHDL"
